Updated Main_Test and Automated_Test
Moved test function implementations to Automated_Test.java
